# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Ember API
  version: "1"
  summary: Desk-display hub for AI agent activity, a Pomodoro timer and friends.
  description: |
    The HTTP contract of the Ember server (`cmd/ember`). Sources post agent
    status and usage, controllers drive the Pomodoro timer, displays read
    state; the macOS app configures everything.

    **Stability.** Operations without `x-internal` are the public v1
    contract: changes are additive only. Operations marked `x-internal: true`
    (the `/v1/device/*` clock proxy, the `…/config` settings and the
    now-playing plumbing) serve Ember.app and may change between releases.

    **Auth.** Reads are open on the LAN. Writes take a bearer token:
    - the master `EMBER_TOKEN` (`ownerToken`) passes every owner route;
    - a scoped client token (`clientToken`, prefix `ekc_`) passes the routes
      whose requirement lists one of its scopes; the `admin` scope passes
      every route that lists any client scope;
    - a knob device token (`deviceToken`, prefix `ekd_`) passes the
      device-self routes and the control routes.

    An `ingest` client token minted with `sources` (not allowed with
    `admin`) is bound to them: `POST`/`DELETE
    /v1/status` and `POST /v1/usage` naming any other `source` (or, for
    usage, none) are 403. Usage stays keyed by tool, so a bound token's usage
    post is also 403 when that tool's stored entry came from a source outside
...

...

...
    every operation.
  license:
    name: MIT
...
servers:
...
components:
  securitySchemes:
    ownerToken:
      type: http
      scheme: bearer
      description: The master `EMBER_TOKEN`. Holds every scope.
    clientToken:
      type: http
      scheme: bearer
      description: |
        A scoped client token (`ekc_…`), minted with `POST /v1/devices`
        `{"kind":"client"}`. Scopes: `ingest`, `control`, `read`, `admin`
        (`admin` satisfies every scope).
    deviceToken:
      type: http
      scheme: bearer
      description: A knob device token (`ekd_…`), minted with `POST /v1/devices` `{"kind":"cinder-knob"}`.
...
paths:
...
  /v1/status:
    post:
      tags: [ingest]
      operationId: postStatus
      summary: Report a session's status
      description: |
        A client token bound to `sources` gets 403 for any other `source`,
        and 400 for a `/` in `tool` or `session`.
      security: [{ownerToken: []}, {clientToken: [ingest]}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusRequest"
            example:
              source: ci
              tool: gha
              session: deploy-42
              state: running
              activity: deploy api
      responses:
        "200":
          description: Stored; returns the aggregated render.
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  render:
                    $ref: "#/components/schemas/Render"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    delete:
      tags: [ingest]
      operationId: deleteStatus
      summary: Drop a session
      description: |
        A client token bound to `sources` gets 403 for any other `source`,
        and 400 for a `/` in `tool` or `session`.
      security: [{ownerToken: []}, {clientToken: [ingest]}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [source, tool, session]
              properties:
                source:
                  type: string
                tool:
                  type: string
                session:
                  type: string
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /v1/clear:
    post:
      tags: [ingest]
      operationId: postClear
      summary: Drop every session
      x-internal: true
      security: [{ownerToken: []}, {clientToken: [admin]}]
      responses:
        "200":
          $ref: "#/components/responses/Object"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /v1/notify:
    post:
      tags: [ingest]
      operationId: postNotify
      summary: Show a one-off notification on the clock
      security: [{ownerToken: []}, {clientToken: [ingest]}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotifyRequest"
            example:
              text: washer done
              color: "#00FF88"
      responses:
        "200":
          description: Sent to the clock.
          content:
            application/json:
              schema:
                type: object
              example:
                ok: true
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "502":
          $ref: "#/components/responses/BadGateway"
  /v1/usage:
    get:
      tags: [read]
      operationId: getUsage
      summary: Latest usage snapshot per tool
      security: []
      responses:
        "200":
          description: Per-tool usage.
          content:
            application/json:
              schema:
                type: object
              examples:
                golden:
                  externalValue: ../cmd/ember/testdata/dashboard/usage.json
    post:
      tags: [ingest]
      operationId: postUsage
      summary: Report a tool's account usage windows
      description: |
        Usage is one entry per tool. A client token bound to `sources` gets
        403 for any other `source`, for none, and when the tool's stored entry
        came from a source outside its list; a tool with no entry yet is
        accepted (the next unbound post overwrites it).
      security: [{ownerToken: []}, {clientToken: [ingest]}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UsageRequest"
            example:
              tool: gemini
              source: dt-mbp
              five_hour:
                used_percent: 42
                resets_at: 1791900000
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /v1/reminders/fire:
    post:
      tags: [ingest]
      operationId: postReminderFire
      summary: Fire a reminder popup on the clock
      security: [{ownerToken: []}, {clientToken: [ingest]}]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReminderFireRequest"
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        "200":
          description: A duplicate of an Idempotency-Key seen in the last 10 minutes; nothing shown.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /v1/reminders/preview:
    get:
      tags: [read]
      operationId: getReminderPreview
      summary: Preview frames of a reminder popup
      security: []
      parameters:
        - name: text
          in: query
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/Object"
  /v1/pomodoro/start:
    post:
      tags: [control]
      operationId: pomodoroStart
      summary: Start a phase
      security: [{ownerToken: []}, {deviceToken: []}, {clientToken: [control]}]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                phase:
                  type: string
                  enum: [focus, short_break, long_break]
                  default: focus
      responses:
        "200":
          description: The timer after the action.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PomodoroStatus"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/pomodoro/pause:
    post:
      tags: [control]
      operationId: pomodoroPause
      summary: Pause the running phase
      security: [{ownerToken: []}, {deviceToken: []}, {clientToken: [control]}]
      responses:
        "200":
          description: The timer after the action.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PomodoroStatus"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/pomodoro/resume:
    post:
      tags: [control]
      operationId: pomodoroResume
      summary: Resume a paused phase
      security: [{ownerToken: []}, {deviceToken: []}, {clientToken: [control]}]
      responses:
        "200":
          description: The timer after the action.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PomodoroStatus"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/pomodoro/stop:
    post:
      tags: [control]
      operationId: pomodoroStop
      summary: Stop the timer
      security: [{ownerToken: []}, {deviceToken: []}, {clientToken: [control]}]
      responses:
        "200":
          description: The timer after the action.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PomodoroStatus"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/pomodoro/skip:
    post:
      tags: [control]
      operationId: pomodoroSkip
      summary: Skip to the next phase
      security: [{ownerToken: []}, {deviceToken: []}, {clientToken: [control]}]
      responses:
        "200":
          description: The timer after the action.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PomodoroStatus"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/pomodoro/state:
    get:
      tags: [read]
      operationId: getPomodoroState
      summary: The timer
      security: []
      responses:
        "200":
          description: The timer now.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PomodoroStatus"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/pomodoro/stats:
    get:
      tags: [read]
      operationId: getPomodoroStats
      summary: Focus statistics
      security: []
      responses:
        "200":
          $ref: "#/components/responses/Object"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/pomodoro/heatmap:
    get:
      tags: [read]
      operationId: getPomodoroHeatmap
      summary: Focus heatmap
      security: []
      responses:
        "200":
          $ref: "#/components/responses/Object"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/pomodoro/workhours:
    get:
      tags: [read]
      operationId: getPomodoroWorkHours
      summary: Focus by hour of day
      security: []
      responses:
        "200":
          $ref: "#/components/responses/Object"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /v1/pomodoro/config:
    get:
      tags: [settings]
      operationId: getPomodoroConfig
      summary: Pomodoro settings
      x-internal: true
      security: [{ownerToken: []}, {clientToken: [read]}]
      responses:
        "200":
          $ref: "#/components/responses/Object"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    put:
      tags: [settings]
      operationId: putPomodoroConfig
      summary: Merge Pomodoro settings
      x-internal: true
      security: [{ownerToken: []}, {clientToken: [admin]}]
      responses:
        "200":
          $ref: "#/components/responses/Object"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /v1/activity/summary:
    get:
      tags: [read]
      operationId: getActivitySummary
      summary: Agent time per tool and source
      security: []
      parameters:
        - name: days
          in: query
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Agent time, waiting excluded.
          content:
            application/json:
              schema:
                type: object
              examples:
                golden:
                  externalValue: ../cmd/ember/testdata/dashboard/activity_summary.json
...
  /v1/clock/stats:
    get:
      tags: [settings]
      operationId: getClockStats
      summary: Clock probe history (memory only)
      x-internal: true
      security: [{ownerToken: []}, {clientToken: [read]}]
      parameters:
        - $ref: "#/components/parameters/StatsRange"
      responses:
        "200":
          description: Probe samples.
          content:
            application/json:
              schema:
                type: object
              examples:
                golden:
                  externalValue: ../cmd/ember/testdata/dashboard/clock_stats.json
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /v1/apps:
    get:
      tags: [settings]
      operationId: getApps
      summary: Per-tool clock visibility
      x-internal: true
      security: [{ownerToken: []}, {clientToken: [read]}]
      responses:
        "200":
          $ref: "#/components/responses/Object"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    put:
      tags: [settings]
      operationId: putApps
      summary: Set per-tool clock visibility
      x-internal: true
      security: [{ownerToken: []}, {clientToken: [admin]}]
      responses:
        "200":
          $ref: "#/components/responses/Object"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /v1/usage/config:
    get:
      tags: [settings]
      operationId: getUsageConfig
      summary: Usage card settings
      x-internal: true
      security: [{ownerToken: []}, {clientToken: [read]}]
      responses:
        "200":
          $ref: "#/components/responses/Object"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    put:
      tags: [settings]
      operationId: putUsageConfig
      summary: Merge usage card settings
      x-internal: true
      security: [{ownerToken: []}, {clientToken: [admin]}]
      responses:
        "200":
          $ref: "#/components/responses/Object"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /v1/display/config:
    get:
      tags: [settings]
      operationId: getDisplayConfig
      summary: Display settings
      x-internal: true
      security: [{ownerToken: []}, {clientToken: [read]}]
      responses:
        "200":
          $ref: "#/components/responses/Object"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    put:
      tags: [settings]
      operationId: putDisplayConfig
      summary: Merge display settings
      x-internal: true
      security: [{ownerToken: []}, {clientToken: [admin]}]
      responses:
        "200":
          $ref: "#/components/responses/Object"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /v1/brightness/config:
    get:
      tags: [settings]
      operationId: getBrightnessConfig
      summary: Brightness schedule settings
      x-internal: true
      security: [{ownerToken: []}, {clientToken: [read]}]
      responses:
        "200":
          $ref: "#/components/responses/Object"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    put:
      tags: [settings]
      operationId: putBrightnessConfig
      summary: Merge brightness schedule settings
      x-internal: true
      security: [{ownerToken: []}, {clientToken: [admin]}]
      responses:
        "200":
          $ref: "#/components/responses/Object"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /v1/quiet/config:
    get:
      tags: [settings]
      operationId: getQuietConfig
      summary: Quiet hours
      x-internal: true
      security: [{ownerToken: []}, {clientToken: [read]}]
      responses:
        "200":
          $ref: "#/components/responses/Object"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    put:
      tags: [settings]
      operationId: putQuietConfig
      summary: Merge quiet hours
      x-internal: true
      security: [{ownerToken: []}, {clientToken: [admin]}]
      responses:
        "200":
          $ref: "#/components/responses/Object"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /v1/weather/config:
    get:
      tags: [settings]
      operationId: getWeatherConfig
      summary: Weather settings (includes the location)
      x-internal: true
      security: [{ownerToken: []}, {clientToken: [admin]}]
      responses:
        "200":
          $ref: "#/components/responses/Object"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    put:
      tags: [settings]
      operationId: putWeatherConfig
      summary: Merge weather settings
      x-internal: true
      security: [{ownerToken: []}, {clientToken: [admin]}]
      responses:
        "200":
          $ref: "#/components/responses/Object"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /v1/meetings/config:
    get:
      tags: [settings]
      operationId: getMeetingsConfig
      summary: Meetings settings (includes the calendar feed)
      x-internal: true
      security: [{ownerToken: []}, {clientToken: [admin]}]
      responses:
        "200":
          $ref: "#/components/responses/Object"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    put:
      tags: [settings]
      operationId: putMeetingsConfig
      summary: Merge meetings settings
      x-internal: true
      security: [{ownerToken: []}, {clientToken: [admin]}]
      responses:
        "200":
          $ref: "#/components/responses/Object"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /v1/devices/self/config:
    get:
      tags: [device-self]
      operationId: deviceSelfConfig
      summary: The calling knob's config
      security: [{deviceToken: []}]
      responses:
        "200":
          description: The config and its version.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckinResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /v1/nowplaying:
    post:
      tags: [nowplaying]
      operationId: postNowPlaying
      summary: A source's player report
      x-internal: true
      security: [{ownerToken: []}, {clientToken: [admin]}]
      responses:
        "200":
          $ref: "#/components/responses/Object"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...